Move app into repo root, remove proxy
Moved contents of app from /osn-discovery-webui/ to /
Reconfigured vite.config.ts to not use a proxy to /api. It caused more
problems and will not be used in production build anyways. Reverted to
using import.meta.env variable insertion.
