+++ src/assets/styles/BucketFile.css
