+++ src/pages/BucketFileListView.tsx
@@ -1,10 +1,10 @@
-import FileColumnHeaders from '../components/FileColumnHeaders.tsx';
-import Loading from '../components/Loading.tsx'
+import FileColumnHeaders from '../../src/components/FileColumnHeaders.tsx';
+import Loading from '../../src/components/Loading.tsx'
 import { useParams } from "react-router-dom"
 import React, { Suspense } from 'react'
 
 const BucketFileListView = () => {
-    const BucketFileList = React.lazy(() => import("../components/BucketFileList.tsx"))
+    const BucketFileList = React.lazy(() => import("../../src/components/BucketFileList.tsx"))
     const routeParams = useParams();
     // Bypass 'string | undefined' typeError when using params
     let key : string = (routeParams['key'] ?? "")
