+++ src/pages/BucketListView.tsx
@@ -1,9 +1,9 @@
-import Loading from '../components/Loading.tsx'
+import Loading from '../../src/components/Loading.tsx'
 import { Suspense, lazy } from 'react'
 
 const BucketListView = () => {
-    const BucketList = lazy(() => import("../components/BucketList.tsx"))
-    const BucketColumnHeaders = lazy(() => import("../components/BucketColumnHeaders.tsx"))
+    const BucketList = lazy(() => import("../../src/components/BucketList.tsx"))
+    const BucketColumnHeaders = lazy(() => import("../../src/components/BucketColumnHeaders.tsx"))
     return (
         <>
             <Suspense fallback={<Loading/>}>
