+++ vite.config.ts
@@ -4,27 +4,27 @@ import react from '@vitejs/plugin-react-swc'
 
 // https://vitejs.dev/config/
 export default defineConfig(({mode}) => {
-  process.env = {...process.env, ...loadEnv(mode, process.cwd())}
+  const env = loadEnv(mode, process.cwd())
   return {
     plugins: [react()],
-    base: './',
     server: {
+      // proxy: {
+      //   "/api": {
+      //     target: env.VITE_API_BASE_URL,
+      //     changeOrigin: true,
+      //     secure: false,
+      //     rewrite: (path) => path.replace(/^\/api/, ""),
+      //   }
+      // },
       watch: {
         usePolling: true,
       },
       host: true, // Needed for Docker Container port mapping to work
       strictPort: true,
       port: 6142,
-      proxy: {
-        '/api': {
-          target: process.env.VITE_API_BASE_URL,
-          changeOrigin: true,
-          rewrite: (path) => path.replace(/^\/api/, ''),
-        }
-      },
-      hmr: {
-        clientPort: 6142
-      },
+      // hmr: {
+      //   clientPort: 6142
+      // },
     }
   }
 })
